Ex 4.3.3.3: Fix ambiguous result in binary search
The provided answer was returning UINT_MAX as the result when the value
was not found in the array.

Since UINT_MAX == (2^32 - 1) on x86_64-unknown-linux-gnu, this means that
if the array has 2^32 or more elements and bsearch() finds the value to
exist in the array at index (2^32 - 1), then it is ambiguous whether the
result of the function call indicates whether the value was found at that
index or whether the value was not found.

To prevent this ambiguity, a postcondition was added ensuring that if the
value is found, then the result must not be the same as when the value is
not found.

This will not pass verification when the "not found" result is UINT_MAX
because `len` can be larger than that, so I increased the "not found"
result to be SIZE_MAX instead of UINT_MAX.

SIZE_MAX is not actually always guaranteed to be the maximum value of
size_t on all platforms, but now we will at least get a verification
error if that happens.

Probably, the most correct way to fix this would be to calculate the
real maximum value of size_t (something like 2^(CHAR_BIT *
sizeof(size_t)) − 1), but that seems a bit too cumbersome.

Alternatively, we could just return `len` when the value is not found,
which is simple and always guaranteed to work, but that does seem a bit
unidiomatic in C...